[Feat] 단일 NIE 조회 API 구현 및 테스트 + 기타 수정사항

+++ src/main/java/com/ada/earthvalley/yomojomo/nie/entities/NieContext.java
@@ -1,6 +1,5 @@
 package com.ada.earthvalley.yomojomo.nie.entities;
 
-import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -17,5 +16,5 @@ import lombok.NoArgsConstructor;
 public class NieContext {
 	@Enumerated(EnumType.STRING)
 	private NieProcessType lastProcess;
-	private long articleCursor;
+	private Integer articleCursor;
 }
